% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction_chip.R
\name{batch_correction_chip}
\alias{batch_correction_chip}
\title{Batch correction}
\usage{
batch_correction_chip(data, reference, value)
}
\arguments{
\item{value}{}
}
\value{

}
\description{
Batch correction
}
