% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_eff.R
\name{cq_eff}
\alias{cq_eff}
\title{cq_eff}
\usage{
cq_eff(qpcr_raw, type = "Cy0")
}
\arguments{
\item{qpcr_raw}{a data frame qpcr raw data with a row for every cycle and a
column for every sample.}

\item{type}{type of method to use for calculating Ct values and efficiencies
among those available for \link[qpcR:pcrbatch]{qpcR::pcrbatch}}
}
\description{
Calculates Cq and efficiency values for every well.
}
