% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_avg_fun.R
\name{ct_avg_fun}
\alias{ct_avg_fun}
\title{Average Cts of technical replicates}
\usage{
ct_avg_fun(ct_data)
}
\arguments{
\item{ct_data}{a data frame with the qPCR input data. It needs to have the following columns: names, cq, efficiency, primers, included.}
}
\value{
a data frame with averaged Ct and standard deviation values for technical replicates.
}
\description{
Calculate the average Ct and standard deviation of technical replicates (rows with the same sample name are considered technical replicates).
}
