% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_avg_plot.R
\name{ct_avg_plot}
\alias{ct_avg_plot}
\title{QC plot of Ct averages of technical replicates}
\usage{
ct_avg_plot(norm_data, max_ct = 35, exp_name = "exp_name")
}
\arguments{
\item{norm_data}{data frame with the normalized qPCR data as generated bu the \code{\link{qpcr_analysis}}}

\item{max_ct}{maximum Ct value considered for quantitative range, a horizontal red line will be drawn in the plot at this value (default = 35)}

\item{exp_name}{experiment name (used as plot title and to name the saved file)}
}
\value{
a pdf file is saved with the corresponding plot
}
\description{
QC plot of Ct averages of technical replicates
}
