% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_sd_plot.R
\name{ct_sd_plot}
\alias{ct_sd_plot}
\title{QC plot of Ct standard deviation of technical replicates}
\usage{
ct_sd_plot(
  norm_data,
  max_sd = sd(c(0, 0.5)),
  max_ct = 35,
  exp_name = "exp_name"
)
}
\arguments{
\item{norm_data}{data frame with the normalized qPCR data as generated bu the \code{\link{qpcr_analysis}}}

\item{max_sd}{maximum standard deviation value between technical replicates to be considered acceptable, a horizontal red line will be drawn in the plot at this value (default = sd(c(0, 0.5)))}

\item{max_ct}{maximum Ct value considered for quantitative range, a vertical red line will be drawn in the plot at this value (default = 35)}

\item{exp_name}{experiment name (used as plot title and to name the saved file)}
}
\value{
a pdf file is saved with the corresponding plot
}
\description{
QC plot of Ct standard deviation of technical replicates
}
