% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct.R
\name{dct}
\alias{dct}
\title{Calculation of dCt values}
\usage{
dct(x, calib, eff.matrix)
}
\arguments{
\item{x}{a row of data from the averaged qPCR data as generated by \code{\link{ct_avg_fun}}, effectively containing the averaged data for a single sample}

\item{calib}{a character string with the name of the sample to be used as reference for dct calculation. This sample will have an
expression value of 1 for all targets and the rest of the samples will have values expressing their relative expression to this sample.}

\item{eff.matrix}{a data.frame containing the efficiency values for each primer pair as generated in \code{\link{qpcr_analysis}}}
}
\value{
the dCt value calculated for the row of data
}
\description{
Function to calculate dCt values according to primer efficiency and a calibrator sample
}
