% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_apply.R
\name{dct_apply}
\alias{dct_apply}
\title{Apply dCt calculation over rows}
\usage{
dct_apply(data, calibsample, eff)
}
\arguments{
\item{data}{a data.frame containing the qPCR data with averaged technical replicates as generated by \code{\link{ct_avg_fun}}.}

\item{calibsample}{a character string with the name of the sample to be used as reference for dct calculation. This sample will have an
expression value of 1 for all targets and the rest of the samples will have values expressing their relative expression to this sample.}

\item{eff}{a data.frame containing the efficiency values for each primer pair as generated in \code{\link{qpcr_analysis}}}
}
\value{
a data.frame containing the qPCR data with an additional column containing the dCt values
}
\description{
A helper function to apply the \code{\link{dct}} function to each row of the data.frame containing the qPCR data
}
