% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_avg_plot.R
\name{eff_avg_plot}
\alias{eff_avg_plot}
\title{QC plot of primer efficiencies}
\usage{
eff_avg_plot(norm_data, effrange = c(1.8, 2), exp_name = "exp_name")
}
\arguments{
\item{norm_data}{data frame with the normalized qPCR data as generated bu the \code{\link{qpcr_analysis}}}

\item{effrange}{a numeric vector with two elements, the lower and higher values for the efficiency range to be considered acceptable}

\item{exp_name}{experiment name (used as plot title and to name the saved file)}
}
\value{
a pdf file is saved with the corresponding plot
}
\description{
A dotplot for QC purposes with efficiency values for each sample and a color scale for the average Ct
}
