% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_well_lightcycler.R
\name{extract_well_lightcycler}
\alias{extract_well_lightcycler}
\title{extract_well_lightcycler}
\usage{
extract_well_lightcycler(dat, i)
}
\arguments{
\item{dat}{fluorescence data for a given cycle.}

\item{i}{well number.}
}
\value{
A list with the well name and fluorescence signal.
}
\description{
Helper function to extract single well data.
}
\keyword{internal}
