% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_names.R
\name{fix_names}
\alias{fix_names}
\title{Fix names}
\usage{
fix_names(x)
}
\arguments{
\item{x}{A character vector of sample names}
}
\value{
A character vector of fixed sample names
}
\description{
Remove well position, dots and asterisk symbols from sample names of chainy output files
}
