% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkg_scatter.R
\name{hkg_scatter}
\alias{hkg_scatter}
\title{Scatterplot of normalization genes}
\usage{
hkg_scatter(dct_data, hkg)
}
\arguments{
\item{hkg}{}
}
\value{

}
\description{
Scatterplot of normalization genes
}
