% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_factor_fun.R
\name{norm_factor_fun}
\alias{norm_factor_fun}
\title{Calculate normalization factor}
\usage{
norm_factor_fun(dct_data, hkg, exp_name)
}
\arguments{
\item{dct_data}{data.frame containing the qPCR data including dCt values as calculated by \code{\link{dct_apply}}}

\item{hkg}{a character vector containing the names of the targets used as normalization genes}

\item{exp_name}{experiment name (used as plot title and to name the saved file)}
}
\value{
a named numeric vector with the normalization factor calculated for each sample. In addition, saves
a PDF containing a scatterplot to asses the correlation between the expression values of normalization genes (if using more than one)
}
\description{
Calculate a normalization factor for each sample using the average of the dct values of all genes used for normalization.
}
