% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_analysis.R
\name{qpcr_analysis}
\alias{qpcr_analysis}
\title{qpcr_analysis}
\usage{
qpcr_analysis(
  ct_data,
  type = "rtqpcr",
  design,
  calibsample,
  hkg,
  exp_name,
  fix_names = FALSE,
  exclude = FALSE,
  save_csv = TRUE
)
}
\arguments{
\item{ct_data}{A data frame with the qPCR input data. It needs to have the following columns: names, cq, efficiency, primers, included.}

\item{type}{Type of analysis, can be \code{"rtqpcr"} or \code{"qchip"}. (NOTE: qchip analysis is not yet implemented)}

\item{design}{A data frame containing the experimental design, at least one expected column called "sample_name" which has to match
the sample names in \code{ct_data} and additional columns for each experimental condition relevant in the design.}

\item{save_csv}{}
}
\value{

}
\description{
qpcr_analysis
}
