% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chainy_dir.R
\name{read_chainy_dir}
\alias{read_chainy_dir}
\title{Read data from chainy output}
\usage{
read_chainy_dir(directory = "./chainy_output/")
}
\arguments{
\item{directory}{directory where the files to read are located}
}
\value{
a data frame with the qPCR data formatted as needed for \code{\link{qpcr_analysis}}
}
\description{
A function to read chainy output files from a specified directory and format them as needed for input in \code{\link{qpcr_analysis}}.
Note that the file name is used as placeholder for the primers column, which works well if every file contains data for a single target.
Otherwise you will have to edit the data to properly annotate the primers before running the \code{\link{qpcr_analysis}} function.
}
