% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_analysis.R
\name{qpcr_analysis}
\alias{qpcr_analysis}
\title{qPCR analysis}
\usage{
qpcr_analysis(
  ct_data,
  design,
  calibsample,
  hkg,
  exp_name,
  fix_names = FALSE,
  exclude = FALSE,
  save_csv = TRUE,
  qc_plots = TRUE
)
}
\arguments{
\item{ct_data}{a data frame with the qPCR input data. It needs to have the following columns: names, cq, efficiency, primers, included.}

\item{design}{a data frame containing the experimental design, at least one expected column called "sample_name" which has to match
the sample names in \code{ct_data} and additional columns for each experimental condition relevant in the design.}

\item{calibsample}{a character string with the name of the sample to be used as reference for dct calculation. This sample will have an
expression value of 1 for all targets and the rest of the samples will have values expressing their relative expression to this sample.}

\item{hkg}{a character vector with the name of the "housekeeping" genes used for normalization.}

\item{exp_name}{a character string with the name of the experiment, which will be used to name the files and objects generated by the function}

\item{fix_names}{logical indicating whether the sample names need to be "fixed" (default = FALSE, only relevant if directly using files generated with chainy to remove well position, dots and asterisks from sample names).}

\item{exclude}{logical indicating whether the logical column "exclude" from the input should be used to filter the data frame (default = FALSE).}

\item{save_csv}{logical indicating whether to save the final analyzed data frame as a csv file.}

\item{qc_plots}{logical indicating whether to generate and save QC plots.}
}
\value{
a data frame with the analyzed qPCR data containing normalized dct values.
}
\description{
A function for the analysis of qPCR data using the "delta Ct" method.
}
